dependencies: split qt out of the ui module
It's a big enough and complicated enough bit of code that it deserves
its own module.
